RDKCOM-3155: REFPLTB-1708 RPI-B 64 bit tdk build failures
Build errors in RPi 64bit BB builds in the tdk-b component:
-Werror=format

Reason for change: REFPLTB-1708 RPI-B 64 bit tdk build failures
Test Procedure: None.
Risks: Low.

(cherry picked from commit 50b33fabe20353c259400489a0c777122f67d67d)

+++ tdkb_lib/src/ssp_pam_bridge.c
@@ -862,7 +862,7 @@ int ssp_CosaDmlDnsGet(char* MethodName, void* cfg)
     else if( !(strcmp(MethodName, "DnsClientServers")) )
     {
        int* temp =cfg;
-       return_status = (int)CosaDmlDnsClientGetServers(pam_handle,&DNSServers);
+       return_status = (int)((long int)CosaDmlDnsClientGetServers(pam_handle,&DNSServers));
        if (return_status != 0)
        {
        printf("ssp_CosaDmlDnsGet: number of servers is %lu\n",DNSServers);
@@ -887,7 +887,7 @@ int ssp_CosaDmlDnsGet(char* MethodName, void* cfg)
     else if( !(strcmp(MethodName, "DnsRelayServers")) )
     {
        int* temp =cfg;
-       return_status = (int)CosaDmlDnsRelayGetServers(pam_handle,&DNSServers);
+       return_status = (int)((long int)CosaDmlDnsRelayGetServers(pam_handle,&DNSServers));
        if (return_status != 0)
        {
        printf("ssp_CosaDmlDnsGet: number of servers is %lu\n",DNSServers);
